fix: 백엔드 API 수정에 따른 Axios API 타입 수정
