--- src/api/index.ts
@@ -1,1 +1,0 @@
-//api관련 함수 및 types
