+++ src/api/modules/auth/index.ts
@@ -1,16 +1,17 @@
 import axiosInstance from 'api/core';
 import { LoginRequest, LoginResponse, RefreshRequest, RefreshResponse } from './types';
+import { CommonResponseReturnType } from '../commonType';
 
 const URL_ROOT = 'auth';
 
 // 로그인
-export const loginFn = async ({ payload }: { payload: LoginRequest }): Promise<LoginResponse> => {
+export const loginFn = async (payload: LoginRequest): Promise<CommonResponseReturnType<LoginResponse>> => {
   const res = await axiosInstance.post(`${URL_ROOT}/login`, payload);
   return res.data;
 };
 
 // 리프레시
-export const refreshFn = async ({ payload }: { payload: RefreshRequest }): Promise<RefreshResponse> => {
+export const refreshFn = async (payload: RefreshRequest): Promise<CommonResponseReturnType<RefreshResponse>> => {
   const res = await axiosInstance.post(`${URL_ROOT}/refresh`, payload);
   return res.data;
 };
