--- src/api/modules/managers/types.ts
@@ -1,15 +1,0 @@
-export type CreateManagerRequest = {
-  accountId: string;
-  name: string;
-  contactNumber: string;
-  email: string;
-}
-
-export type CreateManagerResponse = {
-  id: number;
-  accountId: string;
-  name: string;
-  contactNumber: string;
-  email: string;
-  createdAt: string;
-}
